requirements.txt
dnspython==2.2.1
py3dns==3.2.1
